pyproject.toml
[project]
name = "auditore-eloquence"
version = "1.0.0"
description = "An interactive real-time auditor for professional communication."
authors = [
    { name = "Antares Agentic Environments" }
]
requires-python = ">=3.12"
dependencies = [
    "python-fasthtml",
    "pydantic",
    "google-genai",
    "websockets",
    "uvicorn"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"